feat(processor): creates processor module and listener

+++ arquivo-lib/src/main/java/arquivo/repository/SearchEntityRepository.java
@@ -3,7 +3,5 @@ package arquivo.repository;
 import arquivo.model.SearchEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface SearchEntityRepository extends JpaRepository<SearchEntity, Long> {
+public interface SearchEntityRepository extends JpaRepository<SearchEntity, Integer> {
 }
